% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BWAMEM_Align.R
\name{BWAMEM_Align}
\alias{BWAMEM_Align}
\title{BWAMEM_Align()}
\usage{
BWAMEM_Align(
  Reference_Genome = "No Default",
  Forward_Fastq = "No Default",
  Reverse_Fastq = "No Default",
  Paired_End_Data = TRUE,
  Output_Base_Name = "Bowtie2_Alignment",
  pathtobash = "C:/Program Files/Git/bin/bash.exe",
  pathtoBWA = "bwa",
  pathtoGATK = "./gatk",
  flag = "--rg-id 1 --rg SM:Sample --rg LB:1 --rg PI:400 --rg PL:ILLUMINA"
)
}
\arguments{
\item{Reference_Genome}{The reference genome unto which the sequence data will be mapped (including .fasta or .fna file extension).}

\item{Forward_Fastq}{The single-read or forward paired-end fastq file name (including .fastq file extension)}

\item{Reverse_Fastq}{When Paired_End_Data=TRUE, the reverse paired-end fastq file name (including .fastq file extension). This argument is ignored when Paired_End_Data=FALSE.}

\item{Paired_End_Data}{The input data type, where FALSE corresponds with single-end reads. \link{default=TRUE}}

\item{Output_Base_Name}{The base name for the alignment files and reports. No file extension is required. \link{default=Bowtie2_Alignment}}

\item{pathtobash}{Only for running GATK within a docker (Mainly for windows users): the file path to git bash (bash.exe) when installed in a custom file location. Linux users can alter this directory path to make use of GATK within the docker environment as well.}

\item{pathtoBWA}{Only for running BWA on linux: the file path to the BWA software when a custom installation path has been used. The default is bwa, which assumes gatk is added to your $PATH.}

\item{pathtoGATK}{Only for running GATK on linux: the file path to the GATK software when a custom installation path has been used. The default is ./gatk, which assumes gatk is added to your $PATH.}

\item{flag}{Alignment data is supplied with data information such as read group identifiers (--rg-id), Sample name (--rg SM:), library (rg LB:), Predicted median insert size (--rg PI:) and platform (--rg PL:). \link{default="--rg-id 1 --rg SM:Sample --rg LB:1 --rg PI:400 --rg PL:ILLUMINA"}}
}
\description{
This function allows the user to send commands to BWA-MEM, running either natively on linux and Mac or within a docker for Windows, from within R.
}
\examples{
# Basic single end alignment for windows
BWAMEM_Align(Reference_Genome="ref.fasta",Forward_Fastq="reads.fastq",Paired_End_Data=F,Output_Base_Name="my_alignment")
# Paired end alignment with custom short flag
BWAMEM_Align(Reference_Genome="ref.fasta",Forward_Fastq="fw.fastq",Reverse_Fastq="rv.fastq",Paired_End_Data=T,Output_Base_Name="my_alignment",flag=""--rg-id 1 --rg SM:kip)
}
\keyword{GATK}
\keyword{call}
\keyword{variant}
