% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqTrim.R
\name{FastqTrim}
\alias{FastqTrim}
\title{FastqTrim()}
\usage{
FastqTrim(
  Forward_Fastq = "No Default",
  Reverse_Fastq = "No Default",
  Paired_End_Data = TRUE,
  Adapter_Removal = FALSE,
  Adapters = "No Default",
  Output_Base_Name = "Trimmed_Fastq",
  minLen = 10,
  maxLen = Inf,
  maxN = 30,
  minQ = 10,
  rm.phix = TRUE,
  truncQ = 2,
  trimLeft = 0,
  trimRight = 0,
  maxEE = Inf,
  rm.lowcomplex = 0
)
}
\arguments{
\item{Forward_Fastq}{The single-read or forward paired-end fastq file name (including .fastq file extension)}

\item{Reverse_Fastq}{When Paired_End_Data=TRUE, the reverse paired-end fastq file name (including .fastq file extension). This argument is ignored when Paired_End_Data=FALSE.}

\item{Paired_End_Data}{The input data type, where FALSE corresponds with single-end reads. \link{default=TRUE}}

\item{Adapter_Removal}{Whether to remove adapters sequences defined with Adapters="example.txt" or "sequence". \link{default=FALSE}}

\item{Adapters}{Either an R character string object of forward and reverse adapters or the .txt file of the adapters generated by FastqQual()}

\item{Output_Base_Name}{The base name of the generated trimmed fastq files. No file extension is required. \link{default="Trimmed_Fastq"}}

\item{minLen}{(Optional) The minimum read length allowed. \link{default=20}}

\item{maxLen}{(Optional) The maximum read length allowed. \link{default=Inf, no limit}}

\item{maxN}{(Optional) The maximum number of unknown nucleotides N allowed within a single read. \link{default=30}}

\item{minQ}{(Optional) The minimum base quality score allowed. Reads with one or more bases below this threshold will be discarded. \link{default=10}}

\item{rm.phix}{(Optional) Whether to remove common contraminator sequences recorded in the phiX genome \link{default=TRUE}}

\item{truncQ}{(Optional) The minimum quality score of a base after which a read is truncated \link{default=2}}

\item{trimLeft}{(Optional) The number of nucleotides to remove on the 3' end of all reads. Use this if the first sequence cycles are of low quality. \link{default=0}}

\item{trimRight}{(Optional) The number of nucleotides to remove on the 5' end of all reads. Use this if the last sequence cycles are of low quality. \link{default=0}}

\item{maxEE}{(Optional) The maximum expected error allowed for a single read. EE=sum(10^(-Q/10)). \link{default=Inf, No maximum}}

\item{rm.lowcomplex}{(Optional) The minimum read complexity allowed, based on Shannon Entropy. \link{default=0}}
}
\description{
Based on Dada2 and Rbowtie2, this function allows you to filter both single and paired-end sequence data based on quality, as well as trimming of potential adapters.
}
\examples{
Trim paired-end data on default settings without removing adapters:
FastqTrim(Forward_Fastq="example_fw.fastq",Reverse_Fastq="example_rv.fastq",Paired_End_Data=TRUE,Adapter_Removal=FALSE)
Filter single-end data with custom adapter trimming:
Adapter_String<-"ATGGCATGA"
FastqTrim(Forward_Fastq="example.fastq",Adapter_Removal=TRUE,Adapters=Adapter_String)
}
\keyword{fastq}
\keyword{filter}
\keyword{trim}
